video_controller: 实现获取视频的进度

+++ src/res/result.rs
@@ -3,14 +3,14 @@ use serde::{Deserialize, Serialize};
 // 定义响应数据结构体
 #[derive(Serialize, Deserialize, Debug)]
 pub struct ResponseData<T> {
-    pub code: i32,
+    pub code: i8,
     pub message: String,
     pub data: Option<T>,
 }
 
 // 定义统一响应代码
-pub const SUCCESS_CODE: i32 = 0;
-pub const ERROR_CODE: i32 = -1;
+pub const SUCCESS_CODE: i8 = 0;
+pub const ERROR_CODE: i8 = -1;
 
 // 统一响应结构体的实现
 impl<T> ResponseData<T> {
